+++ src/user.c
@@ -180,7 +180,7 @@ int target_limit_exceeded(Client *client, void *target, const char *name)
 		add_fake_lag(client, 2000); /* lag them up as well */
 
 		flood_limit_exceeded_log(client, "max-concurrent-conversations");
-		sendnumeric(client, ERR_TARGETTOOFAST, name, client->local->nexttarget - TStime());
+		sendnumeric(client, ERR_TARGETTOOFAST, name, (long long)(client->local->nexttarget - TStime()));
 
 		return 1;
 	}
